# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mission Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /runtime-connections:
    get:
      operationId: listRuntimeConnections
      tags: [runtimes]
      summary: List redacted runtime connections
      responses:
        "200":
          description: Runtime connections without secrets or URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RuntimeConnectionSafe"
    post:
      operationId: createRuntimeConnection
      tags: [runtimes]
      summary: Add a server-side runtime connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RuntimeConnectionInput"
      responses:
        "201":
          description: Runtime connection added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeConnectionSafe"
  /runtime-connections/{connectionId}/check:
    post:
      operationId: checkRuntimeConnection
      tags: [runtimes]
      summary: Check a runtime and discover capabilities
      parameters:
        - $ref: "#/components/parameters/ConnectionId"
      responses:
        "200":
          description: Runtime health snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeHealth"
...
  /agents/{agentId}/run:
    get:
      operationId: getAgentRun
      tags: [runtimes]
      summary: Get an agent's current or most recent run
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Agent run, or null when no run exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
        "204":
          description: No run exists
    post:
      operationId: launchAgentRun
      tags: [runtimes]
      summary: Launch a live agent run
      parameters:
        - $ref: "#/components/parameters/AgentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunInput"
      responses:
        "201":
          description: Live run launched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
  /agents/{agentId}/run/pause:
    post:
      operationId: pauseAgentRun
      tags: [runtimes]
      summary: Pause an agent run when supported
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Paused run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
  /agents/{agentId}/run/resume:
    post:
      operationId: resumeAgentRun
      tags: [runtimes]
      summary: Resume an agent run when supported
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Resumed run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
  /agents/{agentId}/run/stop:
    post:
      operationId: stopAgentRun
      tags: [runtimes]
      summary: Safely stop an agent run
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Stopped run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
  /agents/{agentId}/activity:
    get:
      operationId: listAgentActivity
      tags: [runtimes]
      summary: List recent activity for an agent
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Agent runtime activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
  /agents/{agentId}/tool-unlocks:
    get:
      operationId: getAgentToolAccess
      tags: [arena]
      summary: Get a contestant's current Season 0 tool permissions and audit trail
      parameters:
        - $ref: "#/components/parameters/AgentId"
      responses:
        "200":
          description: Tool permission state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentToolAccess"
...
components:
  parameters:
...
    AgentId:
      in: path
      name: agentId
      required: true
      schema:
        type: number
...
    ConnectionId:
      in: path
      name: connectionId
      required: true
      schema:
        type: number
  schemas:
...
    ToolAccess:
      type: object
      required: [key, label, description, available, status, unlockCondition, source, overrideAction, overrideReason, lastChangedAt]
      properties:
        key:
          type: string
        label:
          type: string
        description:
          type: string
        available:
          type: boolean
        status:
          type: string
          enum: [available, locked]
        unlockRound:
          type: [number, "null"]
        unlockCondition:
          type: string
        source:
          type: string
          enum: [free, round, locked, manual]
        overrideAction:
          type: [string, "null"]
          enum: [grant, revoke, null]
        overrideReason:
          type: [string, "null"]
        lastChangedAt:
          type: [string, "null"]
          format: date-time
    ToolUnlockAudit:
      type: object
      required: [id, agentId, toolKey, action, reason, actor, round, createdAt]
      properties:
        id:
          type: number
        agentId:
          type: number
        toolKey:
          type: string
        action:
          type: string
          enum: [grant, revoke]
        reason:
          type: string
        actor:
          type: string
        round:
          type: number
        createdAt:
          type: string
          format: date-time
    AgentToolAccess:
      type: object
      required: [agentId, currentRound, tools, audit]
      properties:
        agentId:
          type: number
        currentRound:
          type: number
        tools:
          type: array
          items:
            $ref: "#/components/schemas/ToolAccess"
        audit:
          type: array
          items:
            $ref: "#/components/schemas/ToolUnlockAudit"
...
    RuntimeConnectionSafe:
      type: object
      required: [id, name, provider, status, capabilities, hasCredentials, lastCheckedAt, errorCode, errorMessage]
      properties:
        id:
          type: number
        name:
          type: string
        provider:
          type: string
          enum: [hermes, openclaw]
        status:
          type: string
          enum: [unknown, healthy, degraded, unhealthy, disabled]
        capabilities:
          type: array
          items:
            type: string
        hasCredentials:
          type: boolean
        lastCheckedAt:
          type: [string, "null"]
          format: date-time
        errorCode:
          type: [string, "null"]
        errorMessage:
          type: [string, "null"]
    RuntimeConnectionInput:
      type: object
      required: [name, provider, endpointUrl]
      properties:
        name:
          type: string
          minLength: 1
        provider:
          type: string
          enum: [hermes, openclaw]
        endpointUrl:
          type: string
        token:
          type: string
          minLength: 1
    RuntimeHealth:
      type: object
      required: [id, name, provider, status, capabilities, lastCheckedAt, errorCode, errorMessage, nextAction]
      properties:
        id:
          type: number
        name:
          type: string
        provider:
          type: string
          enum: [hermes, openclaw]
        status:
          type: string
          enum: [unknown, healthy, degraded, unhealthy, disabled]
        capabilities:
          type: array
          items:
            type: string
        lastCheckedAt:
          type: [string, "null"]
          format: date-time
        errorCode:
          type: [string, "null"]
        errorMessage:
          type: [string, "null"]
        nextAction:
          type: string
        latencyMs:
          type: [number, "null"]
    AgentRunInput:
      type: object
      required: [task]
      properties:
        task:
          type: string
          minLength: 1
        runtimeConnectionId:
          type: number
        tools:
          type: array
          items:
            type: string
    AgentRun:
      type: object
      required: [id, agentId, connectionId, providerRunId, task, allowedTools, status, startedAt, endedAt, lastEventAt, lastEvent, errorCode, errorMessage, supportsPause, supportsResume, supportsStop, createdAt]
      properties:
        id:
          type: number
        agentId:
          type: number
        connectionId:
          type: number
        providerRunId:
          type: [string, "null"]
        task:
          type: string
        allowedTools:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [queued, running, paused, stopping, stopped, completed, failed]
        startedAt:
          type: [string, "null"]
          format: date-time
        endedAt:
          type: [string, "null"]
          format: date-time
        lastEventAt:
          type: [string, "null"]
          format: date-time
        lastEvent:
          type: [string, "null"]
        errorCode:
          type: [string, "null"]
        errorMessage:
          type: [string, "null"]
        supportsPause:
          type: boolean
        supportsResume:
          type: boolean
        supportsStop:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ActivityEvent:
      type: object
      required: [id, agentId, runId, kind, message, createdAt]
      properties:
        id:
          type: number
        agentId:
          type: [number, "null"]
        runId:
          type: [number, "null"]
        kind:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time