# lib/api-zod/src/generated/types/agentRun.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRunStatus } from './agentRunStatus';

export interface AgentRun {
  id: number;
  agentId: number;
  connectionId: number;
  /** @nullable */
  providerRunId: string | null;
  task: string;
  allowedTools: string[];
  status: AgentRunStatus;
  /** @nullable */
  startedAt: Date | null;
  /** @nullable */
  endedAt: Date | null;
  /** @nullable */
  lastEventAt: Date | null;
  /** @nullable */
  lastEvent: string | null;
  /** @nullable */
  errorCode: string | null;
  /** @nullable */
  errorMessage: string | null;
  supportsPause: boolean;
  supportsResume: boolean;
  supportsStop: boolean;
  createdAt: Date;
}