# lib/api-zod/src/generated/types/agentRunInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentRunInput {
  /** @minLength 1 */
  task: string;
  runtimeConnectionId?: number;
  tools?: string[];
}