# lib/api-zod/src/generated/types/agentToolAccess.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolAccess } from './toolAccess';
import type { ToolUnlockAudit } from './toolUnlockAudit';

export interface AgentToolAccess {
  agentId: number;
  currentRound: number;
  tools: ToolAccess[];
  audit: ToolUnlockAudit[];
}