# lib/api-zod/src/generated/types/toolAccess.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolAccessOverrideAction } from './toolAccessOverrideAction';
import type { ToolAccessSource } from './toolAccessSource';
import type { ToolAccessStatus } from './toolAccessStatus';

export interface ToolAccess {
  key: string;
  label: string;
  description: string;
  available: boolean;
  status: ToolAccessStatus;
  /** @nullable */
  unlockRound?: number | null;
  unlockCondition: string;
  source: ToolAccessSource;
  /** @nullable */
  overrideAction: ToolAccessOverrideAction;
  /** @nullable */
  overrideReason: string | null;
  /** @nullable */
  lastChangedAt: Date | null;
}