# lib/api-zod/src/generated/types/toolUnlockAudit.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Mission Control API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolUnlockAuditAction } from './toolUnlockAuditAction';

export interface ToolUnlockAudit {
  id: number;
  agentId: number;
  toolKey: string;
  action: ToolUnlockAuditAction;
  reason: string;
  actor: string;
  round: number;
  createdAt: Date;
}